+++ lib/active_interaction.rb
@@ -20,6 +20,7 @@ require 'active_interaction/concerns/missable'
 require 'active_interaction/concerns/runnable'
 
 require 'active_interaction/grouped_input'
+require 'active_interaction/input'
 require 'active_interaction/inputs'
 
 require 'active_interaction/modules/validation'