+++ spec/active_interaction/filters/abstract_date_time_filter_spec.rb
@@ -3,11 +3,11 @@ require 'spec_helper'
 describe ActiveInteraction::AbstractDateTimeFilter, :filter do
   include_context 'filters'
 
-  describe '#cast' do
+  describe '#process' do
     let(:value) { nil }
 
     it 'raises an error' do
-      expect { filter.send(:cast, value, nil) }.to raise_error NameError
+      expect { filter.process(value, nil) }.to raise_error NameError
     end
   end
 end